feat: Update npm dependencies to include @mui/x-date-pickers and dayjs
Update the npm dependencies in the client/package.json and client/package-lock.json files to include the latest versions of @mui/x-date-pickers and dayjs. This update ensures that the project uses the most up-to-date versions of these libraries, which may include bug fixes, performance improvements, and new features.

Recent repository commits:
- feat: Refactor FormCreateCourses component to use MUI components and react-hook-form
- refactor: Rename FormCourses component to FormCreateCourses
- refactor: Rename FormCourses component to FormCreateCourses
- feat: Add FormCourses component for creating new courses
- feat: Fix typo in Courses page import
- feat: Add validation for creating new courses
- feat: Add Courses model and routes
- feat: Add Contracts page and update navigation
- feat: Add Company page and update navigation